requirements.txt
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
plotly==5.24.1
streamlit==1.41.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
